+++ com.saab.map-streamer/Runtime/Saab.Foundation.Unity/Saab.Foundation.Unity.MapStreamer/MaterialManager.cs
@@ -35,7 +35,6 @@
 //******************************************************************************
 
 // Framework
-using System;
 using System.Collections.Generic;
 
 // Unity